backend/requirements.txt
fastapi
uvicorn
crawl4ai
asyncio
psycopg2-binary